Pathwayのアニメーション保存の修正 Variable選択ダイアログの修正 ProjectExplorerの修正。
git-svn-id: file:///home/moriyoshi/chaperone.e-cell.org/var/opt/subversion/ecell-ide/trunk@3554 fe1fbaf7-ad0a-4b3d-a902-8cb9c273ac6b

+++ ide/PathwayWindow/PathwayWindow/Animation/MassCalculationAnimationItem.cs
@@ -231,8 +231,8 @@ namespace Ecell.IDE.Plugins.PathwayWindow.Animation
         public override void SetAnimationStatus(System.Xml.XmlElement status)
         {
             _autoThreshold = bool.Parse(status.GetAttribute("AutoThreshold"));
-            _thresholdHigh = float.Parse(status.GetAttribute("ThresholdHigh"));
-            _thresholdLow = float.Parse(status.GetAttribute("ThresholdLow"));
+            _thresholdHigh = double.Parse(status.GetAttribute("ThresholdHigh"));
+            _thresholdLow = double.Parse(status.GetAttribute("ThresholdLow"));
             _highEdgeBrush = BrushManager.ParseStringToBrush(status.GetAttribute("HighBrush"));
             _lowEdgeBrush = BrushManager.ParseStringToBrush(status.GetAttribute("LowBrush"));
             _ngEdgeBrush = BrushManager.ParseStringToBrush(status.GetAttribute("NGBrush"));
@@ -366,8 +366,8 @@ namespace Ecell.IDE.Plugins.PathwayWindow.Animation
         {
             if (this.autoThresholdCheckBox.Checked)
                 return;
-            float high = float.Parse(thresholdHigh.Text);
-            float low = float.Parse(thresholdLow.Text);
+            double high = double.Parse(thresholdHigh.Text);
+            double low = double.Parse(thresholdLow.Text);
 
             if (high <= low)
             {
